requirements.txt
Flask==3.0.0
# SQLAlchemy==1.4.36
# Flask-Migrate==3.1.0
# Flask-SQLAlchemy==2.5.1
# Flask-WTF==1.0.0
python-dotenv==1.1.0
Werkzeug==3.0.1
requests==2.32.3
flask-cors==6.0.0
pycryptodome=3.23.0